requirements.txt
python-nmap==0.7.1
simplejson==3.19.3
flask~=3.1.0
customtkinter~=5.2.2